Allow for fractional timezones

+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/products/ProductDetailViewModel.kt
@@ -242,7 +242,7 @@ class ProductDetailViewModel @AssistedInject constructor(
 
     private fun loadParameters() {
         val currencyCode = wooCommerceStore.getSiteSettings(selectedSite.get())?.currencyCode
-        val gmtOffset = selectedSite.get().timezone?.toInt() ?: 0
+        val gmtOffset = selectedSite.get().timezone?.toFloat() ?: 0f
         val (weightUnit, dimensionUnit) = wooCommerceStore.getProductSettings(selectedSite.get())?.let { settings ->
             return@let Pair(settings.weightUnit, settings.dimensionUnit)
         } ?: Pair(null, null)
@@ -333,7 +333,7 @@ class ProductDetailViewModel @AssistedInject constructor(
                 priceWithCurrency = formatCurrency(storedProduct.price, parameters?.currencyCode),
                 salePriceWithCurrency = formatCurrency(storedProduct.salePrice, parameters?.currencyCode),
                 regularPriceWithCurrency = formatCurrency(storedProduct.regularPrice, parameters?.currencyCode),
-                gmtOffset = parameters?.gmtOffset ?: 0
+                gmtOffset = parameters?.gmtOffset ?: 0f
         )
     }
 
@@ -383,7 +383,7 @@ class ProductDetailViewModel @AssistedInject constructor(
         val currencyCode: String?,
         val weightUnit: String?,
         val dimensionUnit: String?,
-        val gmtOffset: Int
+        val gmtOffset: Float
     ) : Parcelable
 
     @Parcelize
@@ -400,7 +400,7 @@ class ProductDetailViewModel @AssistedInject constructor(
         val salePriceWithCurrency: String? = null,
         val regularPriceWithCurrency: String? = null,
         val isProductUpdated: Boolean? = null,
-        val gmtOffset: Int = 0
+        val gmtOffset: Float = 0f
     ) : Parcelable
 
     @Parcelize
